refactor(issuance): address React Doctor findings and remove unused settings UI
