+++ apps/sdp-web/src/app/dashboard/issuance/[tokenId]/asset-profile/tabs/operations-tab.tsx
@@ -1,35 +1,20 @@
 "use client";
 
 import type { Token } from "@sdp/types";
-import {
-  ArrowRightLeft,
-  ChevronDown,
-  Coins,
-  Flame,
-  Lock,
-  type LucideIcon,
-  Pause,
-  Play,
-  ShieldCheck,
-  Snowflake,
-} from "lucide-react";
+import { ChevronDown } from "lucide-react";
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Modal } from "@/components/ui/modal";
 import { useTranslations } from "@/i18n/provider";
 import { TokenDisabledActionTooltip } from "../../token-disabled-action-tooltip";
 import type { TokenOperations } from "../use-token-operations";
+import {
+  getOperationGroups,
+  getOperationLabels,
+  type OperationAction,
+  type OperationRow,
+} from "./operation-rows.model";
 import { OpsActionForms } from "./ops-action-forms";
-
-interface OperationRow {
-  id: string;
-  icon: LucideIcon;
-  title: string;
-  helper: string;
-  actionLabel?: string;
-  onAction: () => void;
-  disabledReason?: string | null;
-}
 
 export function OperationsTab({
   ops,
@@ -41,130 +26,16 @@ export function OperationsTab({
   canManageTokenAdmin: boolean;
 }) {
   const t = useTranslations();
-  const [activeAction, setActiveAction] = useState<
-    "allowlist" | "freeze" | "seize" | "force-burn" | null
-  >(null);
-  const labels = {
-    allowlist: t(
-      token.requiresAllowlist
-        ? "DashboardIssuance.simplified.approvedRecipients"
-        : "DashboardIssuance.simplified.blockedRecipients"
-    ),
-    freeze: t("DashboardIssuance.simplified.freeze"),
-    seize: t("DashboardIssuance.compliance.forceTransfer"),
-    "force-burn": t("DashboardIssuance.compliance.forceBurn"),
-  };
-  const availability = ops.operationAvailability;
-
-  const draftReason = !token.mintAddress ? t("DashboardIssuance.simplified.deployFirst") : null;
-  const isStablecoinDraft = !token.mintAddress && token.template === "stablecoin";
-
-  const supply: OperationRow[] = [
-    {
-      id: "mint",
-      icon: Coins,
-      title: t("DashboardIssuance.management.mintTokens"),
-      helper: t("DashboardIssuance.simplified.mintHint"),
-      actionLabel: t("DashboardIssuance.management.mint"),
-      onAction: () => ops.openFundManagementModal("mint"),
-      disabledReason: availability.mint,
-    },
-    {
-      id: "burn",
-      icon: Flame,
-      title: t("DashboardIssuance.management.burnTokens"),
-      helper: t("DashboardIssuance.simplified.burnHint"),
-      actionLabel: t("DashboardIssuance.management.burn"),
-      onAction: () => ops.openFundManagementModal("burn"),
-      disabledReason: availability.burn,
-    },
-  ];
-  const transfers: OperationRow[] = [];
-  if (ops.showControlList)
-    transfers.push({
-      id: "allowlist",
-      icon: ShieldCheck,
-      title: labels.allowlist,
-      helper: t(
-        token.requiresAllowlist
-          ? "DashboardIssuance.simplified.approvedHint"
-          : "DashboardIssuance.simplified.blockedHint"
-      ),
-      actionLabel: t("DashboardIssuance.simplified.manage"),
-      onAction: () => setActiveAction("allowlist"),
-      // Keep the list readable even when its mutation signer is unavailable.
-    });
-  if (
-    canManageTokenAdmin &&
-    (token.extensions?.pausable || token.status === "paused" || isStablecoinDraft)
-  )
-    transfers.push({
-      id: "pause",
-      icon: token.status === "paused" ? Play : Pause,
-      title: t(
-        token.status === "paused"
-          ? "DashboardIssuance.simplified.resume"
-          : "DashboardIssuance.simplified.pause"
-      ),
-      helper: t(
-        token.status === "paused"
-          ? "DashboardIssuance.simplified.resumeHint"
-          : "DashboardIssuance.simplified.pauseHint"
-      ),
-      onAction: () => ops.handlePause(token.status !== "paused"),
-      actionLabel: t(
-        token.status === "paused"
-          ? "DashboardIssuance.simplified.resumeAction"
-          : "DashboardIssuance.simplified.pauseAction"
-      ),
-      disabledReason: ops.effectivePauseDisabledReason,
-    });
-  if (canManageTokenAdmin && token.isFreezable)
-    transfers.push({
-      id: "freeze",
-      icon: Snowflake,
-      title: labels.freeze,
-      actionLabel: t("DashboardIssuance.simplified.manage"),
-      helper: t("DashboardIssuance.simplified.freezeHint"),
-      onAction: () => setActiveAction("freeze"),
-      disabledReason: ops.effectiveFreezeDisabledReason,
-    });
-  const recovery: OperationRow[] = [];
-  if (canManageTokenAdmin && (token.extensions?.permanentDelegate || isStablecoinDraft))
-    recovery.push(
-      {
-        id: "seize",
-        icon: ArrowRightLeft,
-        title: labels.seize,
-        actionLabel: t("DashboardIssuance.simplified.recoverAction"),
-        helper: t("DashboardIssuance.simplified.forceTransferHint"),
-        onAction: () => setActiveAction("seize"),
-        disabledReason: availability.seize,
-      },
-      {
-        id: "force-burn",
-        icon: Flame,
-        title: labels["force-burn"],
-        actionLabel: t("DashboardIssuance.management.burn"),
-        helper: t("DashboardIssuance.simplified.forceBurnHint"),
-        onAction: () => setActiveAction("force-burn"),
-        disabledReason: availability["force-burn"],
-      }
-    );
-  if (canManageTokenAdmin && ops.lockSupplyRemaining !== null)
-    recovery.push({
-      id: "lock-supply",
-      icon: Lock,
-      title: t("DashboardIssuance.management.lockSupplyTitle"),
-      actionLabel: t("DashboardIssuance.simplified.lockAction"),
-      helper: t("DashboardIssuance.management.lockSupplyHelper"),
-      onAction: ops.openLockSupplyModal,
-      disabledReason: ops.lockSupplyDisabledReason,
-    });
-
-  if (draftReason) {
-    for (const row of [...supply, ...transfers, ...recovery]) row.disabledReason = draftReason;
-  }
+  const [activeAction, setActiveAction] = useState<OperationAction | null>(null);
+  const labels = getOperationLabels(token, t);
+  const { supply, transfers, recovery } = getOperationGroups({
+    ops,
+    token,
+    canManageTokenAdmin,
+    t,
+    onSelect: setActiveAction,
+    labels,
+  });
 
   return (
     <div className="w-full space-y-6">
