+++ apps/sdp-web/src/app/dashboard/issuance/[tokenId]/loading.tsx
@@ -1,4 +1,4 @@
-import { IssuanceDetailSkeleton } from "../issuance-page-skeleton";
+import { IssuanceDetailSkeleton } from "../issuance-detail-skeleton";
 
 export default function IssuanceDetailLoading() {
   return <IssuanceDetailSkeleton />;
