+++ apps/sdp-web/src/app/dashboard/issuance/create/loading.tsx
@@ -1,4 +1,4 @@
-import { IssuanceCreateSkeleton } from "../issuance-page-skeleton";
+import { IssuanceCreateSkeleton } from "../issuance-create-skeleton";
 
 export default function IssuanceCreateLoading() {
   return <IssuanceCreateSkeleton />;
