+++ apps/sdp-web/src/app/dashboard/issuance/legacy-issuance-token-card-skeleton.tsx
@@ -1,0 +1,24 @@
+"use client";
+
+import { SkeletonBlock } from "@/components/ui/skeleton-block";
+
+export function LegacyIssuanceTokenCardSkeleton() {
+  return (
+    <article
+      className="flex min-h-[340px] flex-col rounded-2xl border border-border-default bg-surface-raised p-5"
+      data-loading-card="issuance-token"
+    >
+      <SkeletonBlock className="mb-4 h-14 w-14 rounded-full" />
+      <SkeletonBlock className="h-4 w-16" />
+      <SkeletonBlock className="mt-3 h-8 w-3/4" />
+      <div className="mt-6 space-y-3 rounded-xl border border-border-subtle bg-fill-subtle p-3">
+        <SkeletonBlock className="h-4 w-full" />
+        <SkeletonBlock className="h-4 w-[86%]" />
+        <SkeletonBlock className="h-4 w-[78%]" />
+      </div>
+      <div className="mt-auto pt-3">
+        <SkeletonBlock className="h-11 w-full rounded-[10px]" />
+      </div>
+    </article>
+  );
+}
