+++ apps/sdp-web/src/i18n/ui-copy-baseline.json
@@ -4,17 +4,13 @@
   "src/app/dashboard/integrations/rpc-byok-section.tsx:583:15:https://your-endpoint.example",
   "src/app/dashboard/integrations/rpc-byok-section.tsx:591:71:{API_KEY}",
   "src/app/dashboard/integrations/rpc-byok-section.tsx:740:13:https://your-endpoint.example",
-  "src/app/dashboard/issuance/[tokenId]/asset-profile/asset-profile-header.tsx:114:25:&middot;",
+  "src/app/dashboard/issuance/[tokenId]/asset-profile/asset-profile-header.tsx:118:23:&middot;",
   "src/app/dashboard/issuance/[tokenId]/asset-profile/asset-profile-mapping.ts:175:5:description",
-  "src/app/dashboard/issuance/[tokenId]/asset-profile/tabs/operations-tab.tsx:147:9:force-burn",
-  "src/app/dashboard/issuance/create/issuance-draft-form.tsx:746:17:POST",
-  "src/app/dashboard/issuance/create/issuance-draft-form.tsx:746:29:/v1/issuance/asset-profiles",
-  "src/app/dashboard/issuance/issuance-page-skeleton.tsx:234:11:cap",
-  "src/app/dashboard/issuance/issuance-page-skeleton.tsx:234:11:date",
-  "src/app/dashboard/issuance/issuance-page-skeleton.tsx:234:11:supply",
-  "src/app/dashboard/issuance/issuance-page-skeleton.tsx:251:11:access",
-  "src/app/dashboard/issuance/issuance-page-skeleton.tsx:251:11:controls",
-  "src/app/dashboard/issuance/issuance-page-skeleton.tsx:251:11:frozen",
+  "src/app/dashboard/issuance/[tokenId]/asset-profile/tabs/operation-rows.model.ts:152:9:force-burn",
+  "src/app/dashboard/issuance/create/advanced-settings-editor.tsx:107:15:advanced",
+  "src/app/dashboard/issuance/create/advanced-settings-editor.tsx:99:9:primary",
+  "src/app/dashboard/issuance/create/issuance-draft-form.tsx:739:17:POST",
+  "src/app/dashboard/issuance/create/issuance-draft-form.tsx:739:29:/v1/issuance/asset-profiles",
   "src/app/dashboard/markets/treasury-solutions/treasury-solutions-workspace.tsx:450:11:none",
   "src/app/dashboard/payments/counterparty/counterparty-detail-workspace.tsx:1093:11:details",
   "src/app/dashboard/payments/counterparty/counterparty-detail-workspace.tsx:1093:11:transactions",
